+++ src/main/java/com/team/final8teamproject/board/comment/entity/T_exerciseComment.java
@@ -1,7 +1,6 @@
 package com.team.final8teamproject.board.comment.entity;
 
 import com.team.final8teamproject.board.comment.commentReply.entity.T_exerciseCommentReply;
-import com.team.final8teamproject.share.TimeStamp;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
