+++ src/main/java/com/team/final8teamproject/contact/Comment/entity/ContactComment.java
@@ -2,7 +2,7 @@ package com.team.final8teamproject.contact.Comment.entity;
 
 
 
-import com.team.final8teamproject.user.entity.Timestamped;
+import com.team.final8teamproject.share.exception.TimeStamped;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -28,7 +28,7 @@ import lombok.Setter;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class ContactComment extends Timestamped {
+public class ContactComment extends TimeStamped {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
