+++ src/main/java/com/team/final8teamproject/contact/Comment/service/ContactCommentServiceImpl.java
@@ -19,7 +19,6 @@ public class ContactCommentServiceImpl implements ContactCommentService {
 
   private final ContactCommentRepository contactCommentRepository;
   private final InquiryRepository inquiryRepository;
-  private final UserService userService;
   /**
    * 부모댓글이 있는 경우 . 자식댓글인 경우 (대댓글) 댓글에 댓글을 저장
    * 부모댓글이 없는 경우. 포스트에 댓글을 저장
@@ -29,8 +28,7 @@ public class ContactCommentServiceImpl implements ContactCommentService {
   @Override
   public void saveInquiryComment(Long inquiryId,
       CreateContactCommentRequest createContactCommentRequest,
-      String username) {
-    String nickName = userService.nickNameFindByUsername(username);
+      String username,String nickName) {
     if (!inquiryRepository.existsById(inquiryId)) {
       throw new CustomException(ExceptionStatus.BOARD_NOT_EXIST);
     } else {
