+++ src/main/java/com/team/final8teamproject/contact/entity/Notice.java
@@ -1,9 +1,6 @@
 package com.team.final8teamproject.contact.entity;
 
-import com.team.final8teamproject.contact.dto.NoticeRequest;
-
-import com.team.final8teamproject.contact.dto.UpdateNoticeRequest;
-import com.team.final8teamproject.user.entity.Timestamped;
+import com.team.final8teamproject.share.exception.TimeStamped;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -14,7 +11,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class Notice extends Timestamped {
+public class Notice extends TimeStamped {
 
 
     @Id
